ISAtmosphere.jl

ISAtmosphere.ISAtmosphere — Module

International Standard Atmospheric (ISA) model

The International Standard Atmosphere (ISA) is a static atmospheric model of how the pressure, temperature, density, and viscosity of the Earth's atmosphere change over a wide range of altitudes or elevations. It has been established to provide a common reference for temperature and pressure and consists of tables of values at various altitudes, plus some formulas by which those values were derived.

It is also known as the ICAO Standard Atmosphere, ISA is a standard against which to compare the actual atmosphere at any point and time. The real atmosphere differs from ISA in many ways. Sea level pressure varies from day to day, and there are wide extremes of temperature at all levels.

Only metric units are used.

Source: EUROCONTROL BADA 4 User Manual Chapter 2.2 Atmosphere Model

Source: www.skybrary.aero/index.php/InternationalStandardAtmosphere_(ISA)

Source: en.wikipedia.org/wiki/InternationalStandardAtmosphere

Constants

ISAtmosphere.Hp_trop_m — Constant

Geopotential pressure altitude [m] of Tropopause

ISAtmosphere.R_M²_Ks² — Constant

Real gas constant for air [M²/(Ks²)]

ISAtmosphere.T₀_K — Constant

Standard atmospheric temperature [K] at Mean Sea Level (MSL)

ISAtmosphere.a₀_m_s — Constant

Speed of sound [m/s] at Mean Sea Level (MSL)

ISAtmosphere.g₀_m_s² — Constant

Graviation acceleration [m/s²] at Mean Sea Level (MSL)

ISAtmosphere.p₀_Pa — Constant

Standard atmospheric pressure [Pa] at Mean Sea Level (MSL)

ISAtmosphere.βT∇_K_m — Constant

ISA temperature gradient [K/m] with altitude below the tropopause

ISAtmosphere.κ — Constant

Adiabatic index of air []

ISAtmosphere.ρ₀_kg_m³ — Constant

Standard atmospheric density [kg/m³] at Mean Sea Level (MSL)

Types

ISAtmosphere.AtmosConditions — Type

AtmosConditions(Hp_m, T_K, ΔT_K, p_Pa, ρ_kg_m³, a_m_s)

Immutable STRUCT to keep a set of atmospheric conditions together. This struct can be used to also store an arbitrary set of atmospheric conditions. The function conditions can be used to create the struct.

Functions

ISAtmosphere.Hp_trans_m — Method

Hp_trans_m(Vcas_m_s, M)

Return the transition altitude Hp_trans_m (also called crossover altitude) [m] between a given calibrated airspeed Vcas_m_s [m/s] and a Mach number M.

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-27/28/29

ISAtmosphere.M2Vcas — Method

M2Vcas(M, p_Pa, T_K)

Return the calibrated airspeed Vcas_m_s [m/s] as a function of the Mach number M at pressure level p_Pa [Pa] and with temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-26

ISAtmosphere.M2Vcas — Method

M2Vcas(M, AtmosConditions)

ISAtmosphere.M2Vtas — Method

M2Vtas(M, T_K)

Return true airspeed Vtas_m_s [m/s] as a function of the Mach number M and temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-26

ISAtmosphere.M2Vtas — Method

M2Vtas(M, AtmosConditions)

ISAtmosphere.T_K — Function

T_K(Hp_m [, ΔT_K= 0.0])

Returns the atmospheric temperature [K] at pressure altitude Hp_m [m] and with temperature offset ΔT_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-13/16

ISAtmosphere.Vcas2M — Method

Vcas2M(Vcas_m_s, T_K)

Return the Mach number M as a function of the calibrated airspeed Vcas_m_s [m/s] at pressure level p_Pa [Pa] and with temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-26 (reversed)

ISAtmosphere.Vcas2M — Method

Vcas2M(Vcas_m_s, AtmosConditions)

ISAtmosphere.Vcas2Vtas — Method

Vcas2Vtas(Vcas_m_s, p_Pa, T_K)

Return the true airspeed Vtas_m_s [m/s] as a function of the calibrated airspeed Vcas_m_s [m/s] at pressure level p_Pa [Pa] and with temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-23

ISAtmosphere.Vcas2Vtas — Method

Vcas2Vtas(Vcas_m_s, AtmosConditions)

ISAtmosphere.Vtas2M — Method

Vtas2M(Vtas_m_s, T_K)

Return the Mach number M as a function of the true airspeed Vtas_m_s [m/s] and temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-26 (reversed)

ISAtmosphere.Vtas2M — Method

Vtas2M(Vtas_m_s, AtmosConditions)

ISAtmosphere.Vtas2Vcas — Method

Vtas2Vcas(Vtas_m_s, p_Pa, T_K)

Return calibrated airspeed Vcas_m_s [m/s] as a function of the true airspeed Vtas_m_s [m/s] at pressure level p_Pa [Pa] and with temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-24

ISAtmosphere.Vtas2Vcas — Method

Vtas2Vcas(Vtas_m_s, AtmosConditions)

ISAtmosphere.a_m_s — Method

a_m_s(T_K)

Return the speed of sound [m/s] at the temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-22

ISAtmosphere.conditions — Function

conditions(Hp_m[, ΔT_K = 0.0])

Create AtmosConditions struct with the atmospheric conditions T_K [K], p_Pa [Pa], and ρ_kg_m³ [kg/m³], and the speed of sound a_m_s [m/s] at a given altitude Hp_m [m] and ΔT_K [K] temperature offset.

ISAtmosphere.p_Pa — Function

p_Pa(Hp_m[, ΔT_K = 0.0])

Return the air pressure [Pa] at pressure altitude Hp_m [m] and with temperature offset ΔT_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-18/20

ISAtmosphere.ρ_kg_m³ — Method

ρ_kg_m³(p_Pa, T_K)

Return the air density [kg/m³] at pressure level p_Pa [Pa] and temperature T_K [K].

Source: EUROCONTROL BADA 4 User Manual eq. 2.2-21